Move mockserver as part of mock client to make it configurable
